+++ kikuchipy/io/_io.py
@@ -81,8 +81,8 @@ def load(
 ]:
     """Load an :class:`~kikuchipy.signals.EBSD`,
     :class:`~kikuchipy.signals.EBSDMasterPattern` or
-    :class:`~kikuchipy.signals.ECPMasterPattern` signal from a
-    supported file format.
+    :class:`~kikuchipy.signals.ECPMasterPattern` signal from one of the
+    :ref:`/tutorials/load_save_data.ipynb#Supported-file-formats`.
 
     This function is a modified version of :func:`hyperspy.io.load`.
 
@@ -115,7 +115,7 @@ def load(
     >>> import kikuchipy as kp
     >>> s = kp.load(DATA_DIR + "/patterns.h5")
     >>> s
-    <EBSD, title: patterns My awes0m4 ..., dimensions: (3, 3|60, 60)>
+    <EBSD, title: patterns Scan 1, dimensions: (3, 3|60, 60)>
     """
     if not os.path.isfile(filename):
         is_wildcard = False